chore: move mock endpoints
