+++ Chefs/Chefs.csproj
@@ -254,4 +254,8 @@
 		</Content>
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Client\Data\" />
+	</ItemGroup>
+
 </Project>
