+++ Chefs/Client/Mock/MockNotificationEndpoints.cs
@@ -1,4 +1,4 @@
-namespace Chefs.Client.Data.MockEndpoints;
+namespace Chefs.Client.Mock;
 
 public class MockNotificationEndpoints(string basePath, ISerializer serializer, ILogger<BaseMockEndpoint> logger) : BaseMockEndpoint(serializer, logger)
 {
